# Optimize query count