## Problem

The query log in Laravel uses up memory itself. It stores the query string, bindings, and execution times for each query executed. This adds overhead and affects the benchmark results for time and memory.

## Solution

Listen to the `QueryExecuted` event.

## Results

Test command:
```
<?php

namespace App\Console\Commands;

use Illuminate\Console\Command;
use Illuminate\Support\Facades\DB;

class Experiment extends Command
{
    protected $signature = 'app:experiment {--count=1000 : how many queries to run}';

    public function handle(): void
    {
        $this->line('Running ' . $this->option('count') . ' queries...');

        for ($i = 0; $i < $this->option('count'); $i++) {
            DB::select('select * from users');
        }
    }
}
```

Before:
```
art benchmark app:experiment --count=10000
Running 10000 queries...

⚡  TIME: 513ms   MEM: 4.87MB   SQL: 10000
```
```
art benchmark app:experiment --count=100000
Running 100000 queries...

⚡  TIME: 5.51s   MEM: 38.89MB   SQL: 100000
```

After:
```
$ art benchmark app:experiment --count=10000
Running 10000 queries...

⚡  TIME: 472ms   MEM: 1.03MB   SQL: 10000 
```
```
$ art benchmark app:experiment --count=100000
Running 100000 queries...

⚡  TIME: 5.42s   MEM: 1.03MB   SQL: 100000 
```