@christophrumpel 2025-02-18
Hey, thanks for the PR. That is a good point.
The only disadvantage I see is when the user disabled events for queries in their command, then we wouldn't know any query count.

I wonder if there is another solution like checking directly in Mysql 🤔 What do you think?

````
$this->startQueries = DB::select("SHOW SESSION STATUS LIKE 'Questions'")[0]->Value;
```