@tanerkay 2025-02-18
> The only disadvantage I see is when the user disabled events for queries in their command, then we wouldn't know any query count.

Good point. That would be hard to work around without overriding this and potentially having unwanted side effects.

> I wonder if there is another solution like checking directly in Mysql 🤔 What do you think?
>
> ```
> $this->startQueries = DB::select("SHOW SESSION STATUS LIKE 'Questions'")[0]->Value;
> ```

I like that, though it's a fair bit of work setting this up for each supported database platform (especially testing!)