@christophrumpel 2025-02-18
Gave it another thought. I want the default way to be simple and work for most use cases, which your approach is.

We can still add a flag for the benchmark command to use a different approach of counting if the users decides to disable events.