@tanerkay 2025-02-18
@christophrumpel possible readme update that I didn't think to include with the PR:

> ### Query Count
> The package uses a listener on the Laravel database event `Illuminate\Database\Events\QueryExecuted` to track the number of database queries.
> 
> If events are disabled in your application, queries will not be counted.
> 
> Please note that this only tracks queries executed through Eloquent or the Query Builder. Direct database queries will not be counted. Contributions for improving this functionality are welcome.

or as markdown:

```
### Query Count
The package uses a listener on the Laravel database event `Illuminate\Database\Events\QueryExecuted` to track the number of database queries.

If events are disabled in your application, queries will not be counted.

Please note that this only tracks queries executed through Eloquent or the Query Builder. Direct database queries will not be counted. Contributions for improving this functionality are welcome.
```